+++ src/xyang/ext/__init__.py
@@ -1,0 +1,3 @@
+"""
+Optional extensions to xYang (reference implementations, I-D experiments).
+"""